Filthy but potentially functional hack for selecting the first sentence(s) from the *intro* and not the first /actual/ section
First and foremost, get rid of the initial <div class="mw-parser-output"> so that the "remove all divs" regex doesn't mutilate the intro section.
That way we can ensure that the intro section will remain intact, so that ArticleMetaDescription can choose a snippet from *that* instead, as it's way more likely to be relevant as an overall description of the page in question.

Change-Id: Ib6cf615867da9656ee240037aa94347fcf973378